source/ConsolidatingSensingMatReadings/requirements.txt
keyboard==0.13.5
numpy==1.22.3
pyserial==3.5
